+++ src/dbt_score/dbt_utils.py
@@ -76,6 +76,7 @@ def dbt_ls(select: Iterable[str] | None) -> Iterable[str]:
         "source",
         "snapshot",
         "exposure",
+        "seed",
         "--output",
         "name",
     ]