+++ src/dbt_score/evaluation.py
@@ -65,6 +65,7 @@ class Evaluation:
             self._manifest_loader.sources.values(),
             self._manifest_loader.snapshots.values(),
             self._manifest_loader.exposures.values(),
+            self._manifest_loader.seeds.values(),
         ):
             # type inference on elements from `chain` is wonky
             # and resolves to superclass HasColumnsMixin
@@ -99,5 +100,6 @@ class Evaluation:
             or self._manifest_loader.sources
             or self._manifest_loader.snapshots
             or self._manifest_loader.exposures
+            or self._manifest_loader.seeds
         ):
             self._formatter.project_evaluated(self.project_score)