+++ src/dbt_score/models.py
@@ -154,6 +154,11 @@ class HasColumnsMixin:
         ]
 
 
+# Type annotation for parent references
+ParentType = Union["Model", "Source", "Snapshot", "Seed"]
+ChildType = Union["Model", "Snapshot", "Exposure"]
+
+
 @dataclass
 class Model(HasColumnsMixin):
     """Represents a dbt model.
@@ -180,6 +185,7 @@ class Model(HasColumnsMixin):
         tests: The list of tests attached to the model.
         depends_on: Dictionary of models/sources/macros that the model depends on.
         parents: The list of models, sources, and snapshots this model depends on.
+        children: The list of models and snapshots that depend on this model.
         _raw_values: The raw values of the model (node) in the manifest.
         _raw_test_values: The raw test values of the model (node) in the manifest.
     """
@@ -205,7 +211,8 @@ class Model(HasColumnsMixin):
     tests: list[Test] = field(default_factory=list)
     depends_on: dict[str, list[str]] = field(default_factory=dict)
     constraints: list[Constraint] = field(default_factory=list)
-    parents: list[Union["Model", "Source", "Snapshot"]] = field(default_factory=list)
+    parents: list[ParentType] = field(default_factory=list)
+    children: list[ChildType] = field(default_factory=list)
     _raw_values: dict[str, Any] = field(default_factory=dict)
     _raw_test_values: list[dict[str, Any]] = field(default_factory=list)
 
@@ -245,6 +252,7 @@ class Model(HasColumnsMixin):
                 Constraint.from_raw_values(constraint)
                 for constraint in node_values["constraints"]
             ],
+            parents=[],  # Will be populated later
             _raw_values=node_values,
             _raw_test_values=test_values,
         )
@@ -314,6 +322,7 @@ class Source(HasColumnsMixin):
         patch_path: The yml path of the source definition.
         tags: The list of tags attached to the source table.
         tests: The list of tests attached to the source table.
+        children: The list of models and snapshots that depend on this source.
         _raw_values: The raw values of the source definition in the manifest.
         _raw_test_values: The raw test values of the source definition in the manifest.
     """
@@ -337,6 +346,7 @@ class Source(HasColumnsMixin):
     patch_path: str | None = None
     tags: list[str] = field(default_factory=list)
     tests: list[Test] = field(default_factory=list)
+    children: list[ChildType] = field(default_factory=list)
     _raw_values: dict[str, Any] = field(default_factory=dict)
     _raw_test_values: list[dict[str, Any]] = field(default_factory=list)
 
@@ -419,6 +429,7 @@ class Snapshot(HasColumnsMixin):
         strategy: The strategy of the snapshot.
         unique_key: The unique key of the snapshot.
         parents: The list of models, sources, and snapshots this snapshot depends on.
+        children: The list of models and snapshots that depend on this snapshot.
         _raw_values: The raw values of the snapshot (node) in the manifest.
         _raw_test_values: The raw test values of the snapshot (node) in the manifest.
     """
@@ -443,7 +454,8 @@ class Snapshot(HasColumnsMixin):
     depends_on: dict[str, list[str]] = field(default_factory=dict)
     strategy: str | None = None
     unique_key: list[str] | None = None
-    parents: list[Union["Model", "Source", "Snapshot"]] = field(default_factory=list)
+    parents: list[ParentType] = field(default_factory=list)
+    children: list[ChildType] = field(default_factory=list)
     _raw_values: dict[str, Any] = field(default_factory=dict)
     _raw_test_values: list[dict[str, Any]] = field(default_factory=list)
 
@@ -477,6 +489,7 @@ class Snapshot(HasColumnsMixin):
                 .get("column_name")
             ],
             depends_on=node_values["depends_on"],
+            parents=[],  # Will be populated later
             _raw_values=node_values,
             _raw_test_values=test_values,
         )
@@ -523,7 +536,7 @@ class Exposure:
     meta: dict[str, Any]
     tags: list[str]
     depends_on: dict[str, list[str]] = field(default_factory=dict)
-    parents: list[Union["Model", "Source", "Snapshot"]] = field(default_factory=list)
+    parents: list[ParentType] = field(default_factory=list)
     _raw_values: dict[str, Any] = field(default_factory=dict)
 
     @classmethod
@@ -551,11 +564,91 @@ class Exposure:
         return hash(self.unique_id)
 
 
-Evaluable: TypeAlias = Model | Source | Snapshot | Exposure
+@dataclass
+class Seed(HasColumnsMixin):
+    """Represents a dbt seed.
+
+    Attributes:
+        unique_id: The id of the seed, e.g. `seed.package.seed_name`.
+        name: The name of the seed.
+        relation_name: The relation name of the seed, e.g. `db.schema.seed_name`.
+        description: The full description of the seed.
+        original_file_path: The seed path, e.g. `data/seed_name.csv`.
+        config: The config of the seed.
+        meta: The meta of the seed.
+        columns: The list of columns of the seed.
+        package_name: The package name of the seed.
+        database: The database name of the seed.
+        schema: The schema name of the seed.
+        alias: The alias of the seed.
+        patch_path: The yml path of the seed, e.g. `seeds.yml`.
+        tags: The list of tags attached to the seed.
+        tests: The list of tests attached to the seed.
+        children: The list of models and snapshots that depend on this seed.
+        _raw_values: The raw values of the seed (node) in the manifest.
+        _raw_test_values: The raw test values of the seed (node) in the manifest.
+    """
+
+    unique_id: str
+    name: str
+    relation_name: str
+    description: str
+    original_file_path: str
+    config: dict[str, Any]
+    meta: dict[str, Any]
+    columns: list[Column]
+    package_name: str
+    database: str
+    schema: str
+    alias: str | None = None
+    patch_path: str | None = None
+    tags: list[str] = field(default_factory=list)
+    tests: list[Test] = field(default_factory=list)
+    children: list[ChildType] = field(default_factory=list)
+    _raw_values: dict[str, Any] = field(default_factory=dict)
+    _raw_test_values: list[dict[str, Any]] = field(default_factory=list)
+
+    @classmethod
+    def from_node(
+        cls, node_values: dict[str, Any], test_values: list[dict[str, Any]]
+    ) -> "Seed":
+        """Create a seed object from a node and its tests in the manifest."""
+        return cls(
+            unique_id=node_values["unique_id"],
+            name=node_values["name"],
+            relation_name=node_values["relation_name"],
+            description=node_values["description"],
+            original_file_path=node_values["original_file_path"],
+            config=node_values["config"],
+            meta=node_values["meta"],
+            columns=cls._get_columns(node_values, test_values),
+            package_name=node_values["package_name"],
+            database=node_values["database"],
+            schema=node_values["schema"],
+            alias=node_values["alias"],
+            patch_path=node_values["patch_path"],
+            tags=node_values["tags"],
+            tests=[
+                Test.from_node(test)
+                for test in test_values
+                if not test.get("test_metadata", {})
+                .get("kwargs", {})
+                .get("column_name")
+            ],
+            _raw_values=node_values,
+            _raw_test_values=test_values,
+        )
+
+    def __hash__(self) -> int:
+        """Compute a unique hash for a seed."""
+        return hash(self.unique_id)
+
+
+Evaluable: TypeAlias = Model | Source | Snapshot | Seed | Exposure
 
 
 class ManifestLoader:
-    """Load the models, sources, snapshots, exposures, and tests from the manifest."""
+    """Load the evaluables from the manifest."""
 
     def __init__(self, file_path: Path, select: Iterable[str] | None = None):
         """Initialize the ManifestLoader.
@@ -589,13 +682,15 @@ class ManifestLoader:
         self.sources: dict[str, Source] = {}
         self.snapshots: dict[str, Snapshot] = {}
         self.exposures: dict[str, Exposure] = {}
+        self.seeds: dict[str, Seed] = {}
 
         self._reindex_tests()
         self._load_models()
         self._load_sources()
         self._load_snapshots()
         self._load_exposures()
-        self._populate_parents()
+        self._load_seeds()
+        self._populate_relatives()
 
         if select:
             self._filter_evaluables(select)
@@ -604,6 +699,7 @@ class ManifestLoader:
             len(self.models)
             + len(self.sources)
             + len(self.snapshots)
+            + len(self.seeds)
             + len(self.exposures)
         ) == 0:
             logger.warning("Nothing to evaluate!")
@@ -636,6 +732,13 @@ class ManifestLoader:
                 exposure = Exposure.from_node(node_values)
                 self.exposures[node_id] = exposure
 
+    def _load_seeds(self) -> None:
+        """Load the seeds from the manifest."""
+        for node_id, node_values in self.raw_nodes.items():
+            if node_values.get("resource_type") == "seed":
+                seed = Seed.from_node(node_values, self.tests.get(node_id, []))
+                self.seeds[node_id] = seed
+
     def _reindex_tests(self) -> None:
         """Index tests based on their associated evaluable."""
         for node_values in self.raw_nodes.values():
@@ -652,8 +755,8 @@ class ManifestLoader:
                 ):
                     self.tests[node_unique_id].append(node_values)
 
-    def _populate_parents(self) -> None:
-        """Populate `parents` for all models, snapshots, and exposures."""
+    def _populate_relatives(self) -> None:
+        """Populate `parents` and `children` for all evaluables."""
         for node in (
             list(self.models.values())
             + list(self.snapshots.values())
@@ -662,10 +765,16 @@ class ManifestLoader:
             for parent_id in node.depends_on.get("nodes", []):
                 if parent_id in self.models:
                     node.parents.append(self.models[parent_id])
+                    self.models[parent_id].children.append(node)
                 elif parent_id in self.snapshots:
                     node.parents.append(self.snapshots[parent_id])
+                    self.snapshots[parent_id].children.append(node)
                 elif parent_id in self.sources:
                     node.parents.append(self.sources[parent_id])
+                    self.sources[parent_id].children.append(node)
+                elif parent_id in self.seeds:
+                    node.parents.append(self.seeds[parent_id])
+                    self.seeds[parent_id].children.append(node)
 
     def _filter_evaluables(self, select: Iterable[str]) -> None:
         """Filter evaluables like dbt's --select."""
@@ -685,3 +794,4 @@ class ManifestLoader:
         }
         self.snapshots = {k: s for k, s in self.snapshots.items() if s.name in selected}
         self.exposures = {k: e for k, e in self.exposures.items() if e.name in selected}
+        self.seeds = {k: s for k, s in self.seeds.items() if s.name in selected}