+++ src/dbt_score/rule.py
@@ -14,7 +14,7 @@ from typing import (
     overload,
 )
 
-from dbt_score.models import Evaluable, Exposure, Model, Snapshot, Source
+from dbt_score.models import Evaluable, Exposure, Model, Seed, Snapshot, Source
 from dbt_score.more_itertools import first_true
 from dbt_score.rule_filter import RuleFilter
 
@@ -67,11 +67,13 @@ ModelRuleEvaluationType: TypeAlias = Callable[[Model], RuleViolation | None]
 SourceRuleEvaluationType: TypeAlias = Callable[[Source], RuleViolation | None]
 SnapshotRuleEvaluationType: TypeAlias = Callable[[Snapshot], RuleViolation | None]
 ExposureRuleEvaluationType: TypeAlias = Callable[[Exposure], RuleViolation | None]
+SeedRuleEvaluationType: TypeAlias = Callable[[Seed], RuleViolation | None]
 RuleEvaluationType: TypeAlias = (
     ModelRuleEvaluationType
     | SourceRuleEvaluationType
     | SnapshotRuleEvaluationType
     | ExposureRuleEvaluationType
+    | SeedRuleEvaluationType
 )
 
 
@@ -212,6 +214,11 @@ def rule(__func: SnapshotRuleEvaluationType) -> Type[Rule]:
 
 @overload
 def rule(__func: ExposureRuleEvaluationType) -> Type[Rule]:
+    ...
+
+
+@overload
+def rule(__func: SeedRuleEvaluationType) -> Type[Rule]:
     ...