+++ src/dbt_score/rule_filter.py
@@ -4,18 +4,20 @@ import inspect
 import typing
 from typing import Any, Callable, Type, TypeAlias, cast, overload
 
-from dbt_score.models import Evaluable, Exposure, Model, Snapshot, Source
+from dbt_score.models import Evaluable, Exposure, Model, Seed, Snapshot, Source
 from dbt_score.more_itertools import first_true
 
 ModelFilterEvaluationType: TypeAlias = Callable[[Model], bool]
 SourceFilterEvaluationType: TypeAlias = Callable[[Source], bool]
 SnapshotFilterEvaluationType: TypeAlias = Callable[[Snapshot], bool]
 ExposureFilterEvaluationType: TypeAlias = Callable[[Exposure], bool]
+SeedRuleEvaluationType: TypeAlias = Callable[[Seed], bool]
 FilterEvaluationType: TypeAlias = (
     ModelFilterEvaluationType
     | SourceFilterEvaluationType
     | SnapshotFilterEvaluationType
     | ExposureFilterEvaluationType
+    | SeedRuleEvaluationType
 )
 
 
@@ -91,6 +93,11 @@ def rule_filter(__func: SnapshotFilterEvaluationType) -> Type[RuleFilter]:
 
 @overload
 def rule_filter(__func: ExposureFilterEvaluationType) -> Type[RuleFilter]:
+    ...
+
+
+@overload
+def rule_filter(__func: SeedRuleEvaluationType) -> Type[RuleFilter]:
     ...